Web Development 3/Module 3/Assignment/fast-app/requirements.txt
Flask>=2.3.0
Pillow>=9.5.0
Scrapy>=2.9.0
Werkzeug>=2.3.0